pyproject.toml
[project]
name = "qms-doc-parser"
version = "0.1.0"
description = "Template-aware DOCX parser for QMS documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6.0",
    "PyYAML>=6.0.1",
    "python-docx>=1.1.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]